Consolidate API into single serverless function (fixes Vercel 404)
Vercel's Vite framework integration does not detect individual files
in api/ as serverless functions, returning 404 for all API routes.

Fix: consolidate all endpoints into a single api/index.js handler with
internal URL routing + a rewrite rule to send all /api/* to it.

- Single handler: api/index.js (no Express, no external deps at load)
- All heavy modules lazy-loaded via dynamic import()
- Routes: /api/health, /api/agents, /api/chat, /api/session/:id
- Rewrite: /api/(.*) → /api in vercel.json
