--- api/_lib/cors.js
@@ -1,30 +1,0 @@
-// Simple CORS handler for all API routes
-export function cors(req, res) {
-  res.setHeader('Access-Control-Allow-Origin', '*')
-  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, DELETE, OPTIONS')
-  res.setHeader('Access-Control-Allow-Headers', 'Content-Type')
-
-  if (req.method === 'OPTIONS') {
-    res.status(204).end()
-    return true
-  }
-  return false
-}
-
-// Parse JSON body for POST/DELETE requests
-export async function parseBody(req) {
-  if (req.body) return req.body // Already parsed by Vercel
-
-  return new Promise((resolve, reject) => {
-    let data = ''
-    req.on('data', chunk => { data += chunk })
-    req.on('end', () => {
-      try {
-        resolve(data ? JSON.parse(data) : {})
-      } catch {
-        resolve({})
-      }
-    })
-    req.on('error', reject)
-  })
-}
