+++ vercel.json
@@ -3,9 +3,12 @@
   "outputDirectory": "dist",
   "framework": "vite",
   "functions": {
-    "api/**/*.js": {
+    "api/index.js": {
       "maxDuration": 60,
       "includeFiles": ["agents/**", "agents/workflows/**", "api/_lib/**"]
     }
-  }
+  },
+  "rewrites": [
+    { "source": "/api/(.*)", "destination": "/api" }
+  ]
 }
